Module_B/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.9
mysql-connector-python==8.3.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0
slowapi==0.1.9
pytest==8.2.2
locust==2.24.0
tabulate==0.9.0
